chore: introduce SNT V2 (#1)
This introduces an SNT V2 token implementation. It inherits from
`MiniMeToken` and also comes with an `SNTTestnetController` that can be
used on test networks for allowing minting of tokens by anyone.

This makes the token controller more general and introduces an `open`
property to enable/disable the faucet.

Also adds a `destroyTokens()` function for completeness' sake.